+++ content/browser/renderer_host/cross_process_frame_connector.h
@@ -11,8 +11,8 @@
 #include "base/memory/raw_ptr.h"
 #include "cc/input/touch_action.h"
 #include "components/input/child_frame_input_helper.h"
+#include "content/browser/renderer_host/cross_process_frame_connector_base.h"
 #include "content/common/content_export.h"
-#include "content/public/browser/cross_process_frame_connector_base.h"
 #include "content/public/browser/visibility.h"
 #include "third_party/blink/public/common/frame/frame_visual_properties.h"
 #include "third_party/blink/public/mojom/frame/intrinsic_sizing_info.mojom-forward.h"
